.history/StreamlitApp_20241006195456.py
import os
import json
import pandas as pd
import traceback
import torch
from dotenv import load_dotenv
from src.mcqgenerator.utils import read_file, get_table_data
import streamlit as st
from src.mcqgenerator.MCQgenerator import generate_evaluate_chain
from src.mcqgenerator.logger import logging

