.history/requirements_20241006193817.txt
transformers
accelerate
tensorflow
tokenizers
langchain
streamlit
python-dotenv
PyPDF2
langchain_community
torch
ipywidgets


-e .